+++ src/lib/command/treatment/concat-treatment.js
@@ -1,7 +1,0 @@
-function treat(command, concatCommand) {
-  return concatCommand ? `${command} ${concatCommand}` : command;
-}
-
-module.exports = {
-  treat
-};