+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - `starting-project` input added
 - env variables treated from commands
 - export command treated
+- token not mandatory
 
 ## Bugs: